admin interface

+++ src/admin/rooms/Rooms.tsx
@@ -5,7 +5,7 @@ import {
     FormHelperText, 
     Button,
     IconButton,
-    Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Tooltip,
+    Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Tooltip, Paper,
 } from "@mui/material";
 import AddIcon from '@mui/icons-material/Add';
 import DeleteIcon from '@mui/icons-material/Delete';
@@ -19,9 +19,11 @@ import BinaryDialog from "../../BinaryDialog";
 import { getComparator } from "../../util";
 import Types from "./types/Types";
 import Status from "./status/Status";
+import { paths as api_paths } from "../../api";
+import { TableSkeleton } from "../Table";
 
 function Rooms() {
-    const [rooms, setRooms] = React.useState<Record<string, any>[]>([]);
+    const [rooms, setRooms] = React.useState<Record<string, any>[]|undefined>(undefined);
     const [orderBy, setOrderBy] = React.useState<string>("room_id");
     const [order, setOrder] = React.useState<"asc"|"desc">("asc");
     const [sorted, setSorted] = React.useState<Record<string, any>[]>([]);
@@ -30,7 +32,7 @@ function Rooms() {
     const [del, setDel] = React.useState<Record<string, any>|null>(null);
 
     React.useEffect(() => {
-        fetch("/api/admin/rooms")
+        fetch(api_paths.admin.rooms)
             .then(res => res.json())
             .then(data => {
                 setRooms(data);
@@ -38,13 +40,14 @@ function Rooms() {
     }, []);
 
     React.useEffect(() => {
+        if (rooms === undefined) return;
         const comparator = getComparator(order, orderBy);
         const sorted = [...rooms].sort(comparator);
         setSorted(sorted);
     }, [rooms, order, orderBy]);
 
     React.useEffect(() => {
-        fetch("/api/admin/room_types")
+        fetch(api_paths.admin.room_types)
             .then(res => res.json())
             .then(data => {
                 setRoomTypes(data.reduce((acc: Record<string, any>, cur: Record<string, any>) => {
@@ -55,7 +58,7 @@ function Rooms() {
     }, []);
 
     React.useEffect(() => {
-        fetch("/api/admin/room_status")
+        fetch(api_paths.admin.room_status)
             .then(res => res.json())
             .then(data => {
                 setRoomStatus(data.reduce((acc: Record<string, any>, cur: Record<string, any>) => {
@@ -67,6 +70,7 @@ function Rooms() {
 
     const handleRequestSort = React.useCallback(
         (event: React.MouseEvent<unknown>, property: string) => {
+            if (rooms === undefined) return;
             const isAsc = orderBy === property && order === 'asc';
             setOrder(isAsc ? 'desc' : 'asc');
             setOrderBy(property);
@@ -85,101 +89,115 @@ function Rooms() {
                     active={orderBy === props.field}
                     direction={orderBy === props.field ? order : "asc"}
                     onClick={(e) => { handleRequestSort(e, props.field); }}
-                >{props.label}</TableSortLabel>
+                ><Typography fontWeight="bold">
+                    {props.label}
+                </Typography></TableSortLabel>
             </TableCell>
         );
     };
+
+    const columns = [
+        { field: 'room_id', label: '房间号' },
+        { field: 'name', label: '房间名' },
+        { field: 'capacity', label: '容量' },
+        { field: 'status', label: '状态' },
+        { field: 'type', label: '类型' },
+        { field: 'image', label: '图片', noSort: true },
+        { field: 'action', label: '操作', noSort: true },
+    ];
 
     return(
         <Box>
             <Typography variant="h5" component="h2" gutterBottom>
                 房间管理
             </Typography>
-            <TableContainer sx={{minWidth: 750}}>
-                <Table size="small" stickyHeader>
-                    <TableHead>
-                        <TableRow>
-                            <SortHeadCell field="room_id" label="房间号" />
-                            <SortHeadCell field="name" label="房间名" />
-                            <SortHeadCell field="capacity" label="容量" />
-                            <SortHeadCell field="status" label="状态" />
-                            <SortHeadCell field="type" label="类型" />
-                            <TableCell>图片</TableCell>
-                            <TableCell>操作</TableCell>
-                        </TableRow>
-                    </TableHead>
-                    <TableBody>
-                        {sorted.map((room, i) => {
-                            console.log(room, i);
-                            return (
-                            <TableRow key={i}>
-                                <TableCell>
-                                    {room.room_id}
-                                </TableCell>
-                                <TableCell
-                                    sx={{
-                                        maxWidth: 150,
-                                        overflow: "hidden", textOverflow: "ellipsis",
-                                        whiteSpace: "nowrap"
-                                    }}
-                                >
-                                    {room.name}
-                                </TableCell>
-                                <TableCell>
-                                    {room.capacity}
-                                </TableCell>
-                                <TableCell
-                                    sx={{
-                                        maxWidth: 150,
-                                        overflow: "hidden", textOverflow: "ellipsis",
-                                        whiteSpace: "nowrap"
-                                    }}
-                                >
-                                    <Box display="inline" 
-                                        borderBottom={3} 
-                                        borderColor={statusColors[room.status]}
-                                    >
-                                        {roomStatus[room.status]?.label}
-                                    </Box>
-                                </TableCell>
-                                <TableCell
-                                    sx={{
-                                        maxWidth: 150,
-                                        overflow: "hidden", textOverflow: "ellipsis",
-                                        whiteSpace: "nowrap"
-                                    }}
-                                >
-                                    {roomTypes[room.type]?.label}
-                                </TableCell>
-                                <TableCell>
-                                    <Box display="flex">
-                                            <Box component="img" 
-                                            src={`data:image/png;base64,${room.image}`} 
-                                            alt={room.room_name} 
-                                            maxWidth={50} 
-                                            />
-                                    {room.image? null : <FormHelperText>无图片</FormHelperText>}</Box>
-                                </TableCell>
-                                <TableCell>
-                                    <Tooltip title="删除">
-                                        <IconButton size="small"
-                                            onClick={() => setDel(room)}>
-                                            <DeleteIcon fontSize="small" />
-                                        </IconButton>
-                                    </Tooltip>
-                                    <Tooltip title="编辑">
-                                        <IconButton size="small"
-                                            component={Link} to={`/admin/rooms/edit?room_id=${room.room_id}`}>
-                                            <EditIcon fontSize="small" />
-                                        </IconButton>
-                                    </Tooltip>
-                                </TableCell>
+            {rooms === undefined &&
+            <Paper sx={{ mt: 1, width: '100%', overflow: 'hidden', height: "70vh"}}>
+                <TableSkeleton rowCount={15} columns={columns.map(column => column.label)} />
+            </Paper>}
+            {rooms !== undefined && 
+            <Paper sx={{ mt: 1, width: '100%', overflow: 'hidden' }}>
+                <TableContainer sx={{ minWidth: 750, height: "70vh"}}>
+                    <Table stickyHeader size="small">
+                        <TableHead>
+                            <TableRow>
+                                {columns.map((column, i) => (
+                                    column.noSort ? 
+                                    <TableCell key={i}><Typography fontWeight="bold">{column.label}</Typography></TableCell>:
+                                    <SortHeadCell key={i} field={column.field} label={column.label} />
+                                ))}
                             </TableRow>
-                            );
-                        })}
-                    </TableBody>
-                </Table>
-            </TableContainer>
+                        </TableHead>
+                        <TableBody>
+                            {sorted.map((room, i) => (
+                                <TableRow key={i}>
+                                    <TableCell>
+                                        {room.room_id}
+                                    </TableCell>
+                                    <TableCell
+                                        sx={{
+                                            maxWidth: 150,
+                                            overflow: "hidden", textOverflow: "ellipsis",
+                                            whiteSpace: "nowrap"
+                                        }}
+                                    >
+                                        {room.name}
+                                    </TableCell>
+                                    <TableCell>
+                                        {room.capacity}
+                                    </TableCell>
+                                    <TableCell
+                                        sx={{
+                                            maxWidth: 150,
+                                            overflow: "hidden", textOverflow: "ellipsis",
+                                            whiteSpace: "nowrap"
+                                        }}
+                                    >
+                                        <Box display="inline" 
+                                            borderBottom={3} 
+                                            borderColor={statusColors[room.status]}
+                                        >
+                                            {roomStatus[room.status]?.label}
+                                        </Box>
+                                    </TableCell>
+                                    <TableCell
+                                        sx={{
+                                            maxWidth: 150,
+                                            overflow: "hidden", textOverflow: "ellipsis",
+                                            whiteSpace: "nowrap"
+                                        }}
+                                    >
+                                        {roomTypes[room.type]?.label}
+                                    </TableCell>
+                                    <TableCell>
+                                        <Box display="flex">
+                                                <Box component="img" 
+                                                src={`data:image/png;base64,${room.image}`} 
+                                                alt={room.room_name} 
+                                                maxWidth={50} 
+                                                />
+                                        {room.image? null : <FormHelperText>无图片</FormHelperText>}</Box>
+                                    </TableCell>
+                                    <TableCell>
+                                        <Tooltip title="删除">
+                                            <IconButton size="small"
+                                                onClick={() => setDel(room)}>
+                                                <DeleteIcon fontSize="small" />
+                                            </IconButton>
+                                        </Tooltip>
+                                        <Tooltip title="编辑">
+                                            <IconButton size="small"
+                                                component={Link} to={`/admin/rooms/edit?room_id=${room.room_id}`}>
+                                                <EditIcon fontSize="small" />
+                                            </IconButton>
+                                        </Tooltip>
+                                    </TableCell>
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
+                </TableContainer>
+            </Paper>}
             <Box display="flex" justifyContent="flex-end" pt={2}>
                 <Button fullWidth variant="text" color="primary" startIcon={<AddIcon />}
                     component={Link} to="/admin/rooms/add">
@@ -196,7 +214,7 @@ function Rooms() {
 interface DeleteDialogProps {
     del: Record<string, any>|null;
     setDel: (del: Record<string, any>|null) => void;
-    setRooms: React.Dispatch<React.SetStateAction<Record<string, any>[]>>;
+    setRooms: React.Dispatch<React.SetStateAction<Record<string, any>[]|undefined>>;
 }
 
 const DeleteDialog = ({del, setDel, setRooms}: DeleteDialogProps) => {
@@ -207,26 +225,24 @@ const DeleteDialog = ({del, setDel, setRooms}: DeleteDialogProps) => {
     };
 
     const handleDelete = () => {
-        fetch('/api/admin/rooms', {
+        if (del === null) return;
+        fetch(api_paths.admin.rooms + `/${del.room_id}`, { 
             method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify([{room_id: del?.room_id}])
+            headers: {'Content-Type': 'application/json'},
         })
-        .then(res => {
-            if (res.ok) {
-                showSnackbar({message: "删除成功", severity: "success", duration: 2000});
-                setRooms(old => old.filter(room => room.room_id !== del?.room_id));
+            .then(res => {
+                if (res.ok) {
+                    showSnackbar({message: "删除成功", severity: "success", duration: 2000});
+                    setRooms(old => old?.filter(room => room.room_id !== del.room_id));
+                    handleClose();
+                } else {
+                    throw new Error("删除失败");
+                }
+            })
+            .catch(err => {
+                showSnackbar({message: "删除失败", severity: "error"});
                 handleClose();
-            } else {
-                throw new Error("删除失败");
-            }
-        })
-        .catch(err => {
-            showSnackbar({message: "删除失败", severity: "error"});
-            handleClose();
-        });
+            });
     };
 
     return (
